feat: refactored SignerMessageModal to handle SignerWalletSelector
