+++ src/module/signer/components/display/SignMessageDetails/SignMessageDetails.tsx
@@ -4,6 +4,7 @@ import { ConnectIcon } from "icons";
 import ActionDetailField from "../ActionDetailField/ActionDetailField";
 import Container from "module/common/component/display/Container/Container";
 import { useTranslate } from "module/common/hook/useTranslate";
+import SignerWalletSelector from "module/signer/containers/SignerWalletSelector/SignerWalletSelector";
 
 const SignMessageDetails = ({ receiver, message, metadata }: SignMessageDetailsProps): JSX.Element => {
     const previewProps = metadata ? { dAppPreview: { logoUrl: metadata.logoUrl, Icon: ConnectIcon } } : undefined;
@@ -19,6 +20,7 @@ const SignMessageDetails = ({ receiver, message, metadata }: SignMessageDetailsP
             <Container>
                 <ActionDetailField label={translate("message")} description={message} />
             </Container>
+            <SignerWalletSelector />
         </ActionDetailsScaffold>
     );
 };
